fix: Linux backend robustness sweep
- D-Bus: abstract socket support via SocketAddrExt (the interior-NUL
  path trick never worked — std rejects it), ';'-separated multi-address
  parsing, unknown variant types skipped by signature instead of
  aborting portal response scans
- Window ids validated against the extension's u32 wire type instead of
  silently wrapping
- Portal: listen on the request handle the method reply actually
  returns (pre-0.9 portals), per-request handle tokens, delete the
  portal's saved copy after use (stops littering ~/Pictures)
- Window crops shrink by the off-screen amount instead of shifting
  content in from neighbors
- uinput: mouse devices no longer claim all 256 keyboard keys (can get
  classified as keyboards); keybit ioctl results checked; 'ctrl++'
  presses the literal plus key; shifted chars in combos press shift
  (ctrl+% is ctrl+shift+5, not ctrl+5); compile_error on unsupported
  architectures; serialize portal-using integration tests

+++ src/platform/linux/windows.rs
@@ -24,11 +24,11 @@ pub fn list_windows() -> Result<String, String> {
     ]))
 }
 
-pub fn raise_window(id: u64) -> Result<String, String> {
+pub fn raise_window(id: u32) -> Result<String, String> {
     let mut conn = DbusConnection::connect()?;
 
     let mut body = MarshalBuffer::new();
-    body.write_u32(id as u32);
+    body.write_u32(id);
 
     conn.call_method(
         DEST, PATH, IFACE,
